# backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: AI Services Platform API
  version: 2.0.0
  description: APIs for AI-powered services including Virtual Assistant, IDP, Computer Vision, Authentication, and more.
  contact:
    name: AI Services Platform Support
    email: support@aiservicesplatform.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/products:
    get:
      summary: Get all available products
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product (Admin only)
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      summary: Get product by ID
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    $ref: '#/components/schemas/Product'
    put:
      summary: Update product (Admin only)
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    $ref: '#/components/schemas/Product'
    delete:
      summary: Delete product (Admin only)
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/products/{id}/status:
    patch:
      summary: Update product status (Admin only)
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
              required: [isActive]
      responses:
        '200':
          description: Product status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    $ref: '#/components/schemas/Product'

  # Product Configuration Endpoints
  /api/product-configurations:
    get:
      summary: Get all product configurations for current tenant
      tags: [Product Configurations]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of product configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  configurations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductConfiguration'
        '401':
          description: Unauthorized
    post:
      summary: Create or update product configuration for a specific product
      tags: [Product Configurations]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  description: The product ID to configure
                configuration:
                  type: object
                  description: Product-specific configuration settings
              required: [productId, configuration]
      responses:
        '200':
          description: Configuration saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  configuration:
                    $ref: '#/components/schemas/ProductConfiguration'
        '401':
          description: Unauthorized
  /api/product-configurations/{productId}:
    get:
      summary: Get configuration for a specific product
      tags: [Product Configurations]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
          description: The product ID to get configuration for
      responses:
        '200':
          description: Product configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  configuration:
                    $ref: '#/components/schemas/ProductConfiguration'
        '404':
          description: Configuration not found
        '401':
          description: Unauthorized
    delete:
      summary: Delete configuration for a specific product (Admin only)
      tags: [Product Configurations]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '403':
          description: Admin access required
        '401':
          description: Unauthorized

  # Assistant Channels Endpoints
...
  /api/subscriptions/product/{productId}:
    get:
      summary: Get subscription for a specific product
      tags: [Subscriptions]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  subscription:
                    $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [Virtual Assistant, IDP, Computer Vision]
        features:
          type: array
          items:
            type: string
        pricing:
          type: object
          properties:
            tiers:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
                  billingPeriod:
                    type: string
                  features:
                    type: array
                    items:
                      type: string
        defaultPromptTemplateId:
          type: string
          description: Reference to the default prompt template for this product
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [Virtual Assistant, IDP, Computer Vision]
        features:
          type: array
          items:
            type: string
        pricing:
          type: object
        defaultPromptTemplateId:
          type: string
        isActive:
          type: boolean
      required: [name, description, category]
    ProductConfiguration:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: string
          description: The tenant this configuration belongs to
        productId:
          type: string
          description: The product this configuration is for
        userId:
          type: string
          description: User who created/updated the configuration
        configuration:
          type: object
          description: Product-specific configuration key-value pairs
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, inactive]
          description: Configuration status
...
    Subscription:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: string
        customerId:
          type: string
        productId:
          type: string
        tierId:
          type: string
          description: Product tier ID
        status:
          type: string
          enum: [active, inactive, cancelled, suspended]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        autoRenew:
          type: boolean
        usage:
          type: object
          properties:
            sttSeconds:
              type: number
            ttsCharacters:
              type: number
            llmTokensIn:
              type: integer
            llmTokensOut:
              type: integer
            totalCost:
              type: number
            lastUpdated:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...